Refactor main class
